Set StartupNotify to false in the desktop entry file.
When lauching XTerm and Guake, the mouse cursor would wrongly
display a busy indicator for 15 seconds after the applications
had been opened.

The Desktop Entry Specification states that StartupNotify should
only be true if it's known that the application being started will
send a "remove" message--as defined by the Startup Notification
Protocol. A "remove" message ends the startup sequence indicating
that the application is ready to use (systemd services indicate
this condition with a "READY=1" notificaton).

XTerm and Guake do not implement the Startup Notification Protocol
so we are using "StartupNotify=false" until this can be handled for
all supported terminal emulators.