Add filterwarnings() for `gnome-terminal-server` ResourceWarning.
It is possible to call the process with os.spawn* methods, but for
some reason, it only works when `gnome-terminal-server` is passed as
the command and again as the first argument. For example:

    os.spawnl(
        os.P_NOWAIT,                           # <- mode
        '/usr/libexec/gnome-terminal-server',  # <- path (command)
        '/usr/libexec/gnome-terminal-server',  # <- arg 1
        '--app-id',                            # <- arg 2
        'com.github.shawnbrown.EnvLauncher',   # <- arg 3
    )

While this works and doesn't trigger a ResourceWarning, I'm not sure
WHY it works so I don't know if it's safe or reliable. So even though
the current implementation triggers a warning, the process DOES end
when it's no longer needed and the behavior operates according to
known, documented behaviors.