Update `rcfile` and `banner` normalization.
Also remove tests for bogus assignment and add type hints instead.

This change moves the normalization from the property setters into
the getters. Doing this makes sure we are checking the data as it
comes in to the application rather than when it leaves the app.

This also means that we are trusting data generated from within
the app. And the type hints have been added to help check that
the implementation is correct.